- fixed grenade and glauncher physics on big maps
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@3488 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Tactical/LOS.cpp
@@ -5323,7 +5323,7 @@ INT32	CheckForCollision( FLOAT dX, FLOAT dY, FLOAT dZ, FLOAT dDeltaX, FLOAT dDel
 	FLOAT						dTargetZMax;
 	BOOLEAN					fIntended;
 
-	INT16						sX, sY, sZ;
+	INT32						sX, sY, sZ;
 
 	FLOAT						dOldZUnits, dZUnits;
 
@@ -5331,9 +5331,9 @@ INT32	CheckForCollision( FLOAT dX, FLOAT dY, FLOAT dZ, FLOAT dDeltaX, FLOAT dDel
 	INT32						iCurrCubesZ;
 
 
-	sX = (INT16)( dX / CELL_X_SIZE );
-	sY = (INT16)( dY / CELL_Y_SIZE );
-	sZ = (INT16)dZ;
+	sX = (INT32)( dX / CELL_X_SIZE );
+	sY = (INT32)( dY / CELL_Y_SIZE );
+	sZ = (INT32)dZ;
 
 	// Check if gridno is in bounds....
 	if ( !GridNoOnVisibleWorldTile( sX + sY * WORLD_COLS ) )
